RT#36864 Fake select on filehandles on win32
Tests were failing as a result of lack of support for using select with
filehandles on win32 (see perlport). Since IO::Select is used
pervasively in Stream.pm, it was easier to just make can_read return all
handles known to the object.  This works fine since OpenFile is used
only by the test suite (for better or worse).